Refactored LinterRulesProvider
