+++ src/Bicep.Core/Analyzers/Linter/ILinterRulesProvider.cs
@@ -1,12 +1,16 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using System;
+using System.Collections.Generic;
 using System.Collections.Immutable;
 
-namespace Bicep.LanguageServer.Providers
+namespace Bicep.Core.Analyzers.Linter
 {
     public interface ILinterRulesProvider
     {
         ImmutableDictionary<string, string> GetLinterRules();
+
+        IEnumerable<Type> GetRuleTypes();
     }
 }
