+++ src/Bicep.LangServer.UnitTests/BicepCompilationManagerHelper.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using Bicep.Core;
+using Bicep.Core.Analyzers.Linter;
 using Bicep.Core.Configuration;
 using Bicep.Core.FileSystem;
 using Bicep.Core.Registry;
@@ -95,5 +96,5 @@ namespace Bicep.LangServer.UnitTests
 
             return scheduler;
         }
-      }
+    }
 }
