+++ src/Bicep.LangServer/Handlers/BicepTextDocumentSyncHandler.cs
@@ -5,10 +5,10 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Bicep.Core;
+using Bicep.Core.Analyzers.Linter;
 using Bicep.Core.Configuration;
 using Bicep.LanguageServer.CompilationManager;
 using Bicep.LanguageServer.Configuration;
-using Bicep.LanguageServer.Providers;
 using Bicep.LanguageServer.Telemetry;
 using Bicep.LanguageServer.Utils;
 using MediatR;
